v0.2b2

+++ release/clean.sh
@@ -9,7 +9,10 @@ DOA_CONF_DIR=/var/lib/madeira/opsagent
 DOA_LOG_DIR=/var/log/madeira
 DOA_TMP_ROOT=/tmp/opsagent
 DOA_ROOT_DIR=/opt/madeira
-DOA_REMOTE=https://s3.amazonaws.com/visualops
+DOA_BASE_REMOTE=https://s3.amazonaws.com/visualops
+#DOA_VERSION=""
+#DOA_REMOTE="${DOA_BASE_REMOTE}/${DOA_VERSION}"
+EXIT=0
 
 # DEBUG
 DOA_SALT_DIR=/opt/madeira/bootstrap
@@ -30,7 +33,7 @@ else
     exit 1
 fi
 
-(crontab -l | grep -v ${DOA_CONF_DIR}/cron.sh) > /tmp/opsagent.crontab
+(crontab -l | grep -v ${DOA_CONF_DIR}) > /tmp/opsagent.crontab
 crontab -r
 cat /tmp/opsagent.crontab | crontab
 #mv -f /tmp/opsagent.crontab /etc/crontab
@@ -46,41 +49,48 @@ rm -rf ${DOA_ROOT_DIR}
 rm -rf ${DOA_TMP_ROOT}*
 
 if [ "$1" = "reinstall" ]; then
-    curl -sSL ${DOA_REMOTE}/userdata.sh | bash
-fi
-
-if [ "$2" = "debug" ]; then
-    (crontab -l | grep -v ${DOA_CONF_DIR}/cron.sh) > /tmp/opsagent.crontab
-    crontab -r
-    cat /tmp/opsagent.crontab | crontab
-#    (crontab -l | grep -v ${DOA_CONF_DIR}/cron.sh) > ${DOA_TMP_ROOT}.crontab
-#    cp -f ${DOA_TMP_ROOT}.crontab /etc/crontab
-    bash ${DOA_CONF_DIR}/cron.sh
-    service opsagentd stop
-    for id in `ps aux | grep opsagent | sed -e 's/  */ /g' | cut -d ' ' -f 2`; do
-        kill -9 $id
-    done
-    rm -rf ${DOA_TMP_ROOT}*
-    if [ $(which chkconfig) ]; then
-        chkconfig --del opsagentd
-    elif [ $(which update-rc.d) ]; then
-        update-rc.d opsagentd disable
-    else
-        echo "Fatal: no service manager"
-        exit 1
+    curl -sSL "http://169.254.169.254/latest/user-data" -o /tmp/userdata.sh
+    if [ $? -ne 0 ]; then
+        curl -sSL ${DOA_REMOTE}/userdata.sh -o /tmp/userdata.sh
     fi
 
-    cd $DOA_SALT_DIR
-    rm -rf $DOA_SALT
-    git clone $DOA_SALT_REPO $DOA_SALT
-    cd $DOA_SALT
-    git checkout $DOA_SALT_BRANCH
-    rm -f ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/salt
-    ln -s ${DOA_BOOT_DIR}/${DOA_SALT}/sources/salt ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/salt
-    rm -f ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/opsagent/state/adaptor.py
-    ln -s ${DOA_BOOT_DIR}/${DOA_SALT}/sources/adaptor.py ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/opsagent/state/adaptor.py
+    bash /tmp/userdata.sh
+    EXIT=$?
+    if [ $EXIT -eq 0 ] && [ "$2" = "debug" ]; then
+        (crontab -l | grep -v ${DOA_CONF_DIR}) > /tmp/opsagent.crontab
+        crontab -r
+        cat /tmp/opsagent.crontab | crontab
+        #(crontab -l | grep -v ${DOA_CONF_DIR}/cron.sh) > ${DOA_TMP_ROOT}.crontab
+        #cp -f ${DOA_TMP_ROOT}.crontab /etc/crontab
+        bash ${DOA_CONF_DIR}/cron.sh
+        EXIT=$?
+
+        service opsagentd stop
+        for id in `ps aux | grep opsagent | sed -e 's/  */ /g' | cut -d ' ' -f 2`; do
+            kill -9 $id
+        done
+        rm -rf ${DOA_TMP_ROOT}*
+        if [ $(which chkconfig) ]; then
+            chkconfig --del opsagentd
+        elif [ $(which update-rc.d) ]; then
+            update-rc.d opsagentd disable
+        else
+            echo "Fatal: no service manager"
+            exit 1
+        fi
+
+        cd $DOA_SALT_DIR
+        rm -rf $DOA_SALT
+        git clone $DOA_SALT_REPO $DOA_SALT
+        cd $DOA_SALT
+        git checkout $DOA_SALT_BRANCH
+        rm -f ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/salt
+        ln -s ${DOA_BOOT_DIR}/${DOA_SALT}/sources/salt ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/salt
+        rm -f ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/opsagent/state/adaptor.py
+        ln -s ${DOA_BOOT_DIR}/${DOA_SALT}/sources/adaptor.py ${DOA_ENV_DIR}/lib/${D_PYTHON}/site-packages/opsagent/state/adaptor.py
+    fi
 fi
 
 
-exit 0
+exit $EXIT
 # EOF
